+++ packages/backend-app-api/package.json
@@ -71,7 +71,7 @@
     "minimist": "^1.2.5",
     "morgan": "^1.10.0",
     "node-cache": "^5.1.2",
-    "node-fetch": "^3.3.2",
+    "node-fetch": "^2.6.7",
     "node-forge": "^1.3.1",
     "selfsigned": "^2.0.0",
     "stoppable": "^1.1.0",
