+++ plugins/discovery-backend/api-report.md
@@ -1,0 +1,11 @@
+## API Report File for "@backstage/plugin-discovery-backend"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
+
+// @public (undocumented)
+const discoveryPlugin: () => BackendFeature;
+export default discoveryPlugin;
+```
