+++ plugins/discovery-backend/catalog-info.yaml
@@ -1,0 +1,10 @@
+apiVersion: backstage.io/v1alpha1
+kind: Component
+metadata:
+  name: backstage-plugin-discovery-backend
+  title: '@backstage/plugin-discovery-backend'
+  description: Discovery plugin for HTTP based discovery.
+spec:
+  lifecycle: experimental
+  type: backstage-backend-plugin
+  owner: maintainers
