+++ plugins/discovery-backend/package.json
@@ -12,7 +12,7 @@
     "types": "dist/index.d.ts"
   },
   "backstage": {
-    "role": "backend-plugin-module"
+    "role": "backend-plugin"
   },
   "scripts": {
     "start": "backstage-cli package start",
